API refactor (WIP)
